pyproject.toml
[tool.poetry]
name = "seibox"
version = "0.1.0"
description = "Safety Evals in a Box - A reproducible harness to evaluate and improve LLM safety"
authors = ["Safety Evals Team"]
readme = "README.md"
packages = [{include = "seibox"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.0"
pydantic = "^2.6.0"
pyyaml = "^6.0"
click = "^8.1.7"
pytest = "^8.0.0"
streamlit = "^1.31.0"
orjson = "^3.9.13"
python-dotenv = "^1.0.1"
tqdm = "^4.66.2"
rich = "^13.7.0"
xxhash = "^3.4.1"
jinja2 = "^3.1.3"
openai = "^1.12.0"
anthropic = "^0.64.0"
tiktoken = "^0.8.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.0"
black = "^24.1.0"
mypy = "^1.8.0"
pytest-cov = "^6.2.1"
types-pyyaml = "^6.0.12"
types-tqdm = "^4.66.0"

[tool.poetry.scripts]
seibox = "seibox.__main__:cli"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP"]
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"